osd: EC Optimizations fix routing of requests to non-zero shard id
Pools with EC optimizations can use pg_temp to modify the selection
of the primary. When this happens the clients route request to the
correct OSD, but wrong shard which causes hung I/Os or misdirected
I/Os.

Fix Objecter to select the correct shard when sending requests to
EC optimized pools. Fix OSD to modify the shard when receving
requests from legacy clients.

Add new unittests to test new functions for remapping the shard id.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>