build-integration-branch: allow setting git trailer on final commit
After the last commit is made, provide a simple mechanism for adding
git trailers to the commit message. The git trailers [1] are metadata
that tools may make use of. In particular, we add a few of the
trailers documented by ceph-build here [2] as well as allowing
for arbitrary trailers for future changes (before this code can
be updated), advanced trailer, or other unrelated purposes.

[1] https://www.alchemists.io/articles/git_trailers
[2] https://github.com/ceph/ceph-build/tree/main/ceph-trigger-build

Signed-off-by: John Mulligan <jmulligan@redhat.com>